Changed double to int

+++ src/Main.java
@@ -2,15 +2,15 @@ import java.util.Random;
 
 public class Main {
     public static void main(String[] args) {
-        String accessOutside1 = accessForWalking(21.0, 0);
+        String accessOutside1 = accessForWalking(21, 0);
         System.out.println(accessOutside1);
-        String accessOutside2 = accessForWalking(19.0, 25.0);
+        String accessOutside2 = accessForWalking(19, 25);
         System.out.println(accessOutside2);
-        String accessOutside3 = accessForWalking(46.0, -9.0);
+        String accessOutside3 = accessForWalking(46, -9);
         System.out.println(accessOutside3);
-        String accessOutside4 = accessForWalking(13.0, -10.0);
+        String accessOutside4 = accessForWalking(13, -10);
         System.out.println(accessOutside4);
-        String accessOutside5 = accessForWalking(19.0, -21.0);
+        String accessOutside5 = accessForWalking(19, -21);
         System.out.println(accessOutside5);
         int age = generateRandomAge();
         System.out.println(age);
@@ -18,7 +18,7 @@ public class Main {
         System.out.println(randomAgeAccessOutside);
 
     }
-    public static String accessForWalking (double ageOfMan, double temperatureOutside){
+    public static String accessForWalking (int ageOfMan, int temperatureOutside){ /*Changed double to int*/
         String walking = "Можно идти гулять";
         if (ageOfMan >= 20 && ageOfMan <= 45 && temperatureOutside >= -20 && temperatureOutside <= 30) {
         } else if (ageOfMan < 20 && temperatureOutside >= 0 && temperatureOutside <= 28) {
